WebDevelopment/IP-VEA/src/bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = "Bot for Messaging applicatioin"
authors = ["Yóunǎi <c0d3r.nodiru.gaji@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
aiogram = "^3.10.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
